# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "all admin",
                "description": "Admin",
                "operationId": "27337c13def5395d38df0502318680d2",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "join"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "john"
                                        },
                                        "email": {
                                            "type": "email",
                                            "example": "a@a.com"
                                        },
                                        "password": {
                                            "type": "password",
                                            "example": "01020304"
                                        },
                                        "phone": {
                                            "type": "number",
                                            "example": "1234567"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post a new data",
                "description": "Post new admin",
                "operationId": "cf794e8531b75966d24dd460b4e731ec",
                "requestBody": {
                    "description": "Admin",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "join"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "john"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "a@a.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "01020304"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": "1234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/admin/{admin}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "admin",
                "description": "Returns project data",
                "operationId": "18e66cc83a15cbeb49037ee8e011383c",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post a new data",
                "description": "update  admin  data",
                "operationId": "8ea23aae17ff66bda72308d46c0fe20c",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "description": "Id of admin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass admin   credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "integer",
                                        "format": "foreign",
                                        "example": "john"
                                    },
                                    "surname": {
                                        "type": "integer",
                                        "format": "foreign",
                                        "example": "bb"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "a@agmail.com"
                                    },
                                    "password": {
                                        "type": "passsword",
                                        "format": "text",
                                        "example": "12345680"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "4431063"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteAdmin",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "description": "Admin id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "api"
        }
    }
}